requirements.txt
python-telegram-bot==12.8
psycopg2-binary
matplotlib~=3.5.0
numpy~=1.21.4
scipy~=1.7.3